Moved provenance_constants.py and provenance_profile.py to cwlprov/
