+++ cwltool/cwlprov/ro.py
@@ -37,7 +37,7 @@ from schema_salad.utils import json_dumps
 # from typing_extensions import TYPE_CHECKING, TypedDict
 
 from ..loghandler import _logger
-from ..provenance_constants import (
+from .provenance_constants import (
     ACCOUNT_UUID,
     CWLPROV,
     CWLPROV_VERSION,
