+++ cwltool/workflow.py
@@ -29,7 +29,7 @@ from .errors import WorkflowException
 from .load_tool import load_tool
 from .loghandler import _logger
 from .process import Process, get_overrides, shortname
-from .provenance_profile import ProvenanceProfile
+from .cwlprov.provenance_profile import ProvenanceProfile
 from .utils import (
     CWLObjectType,
     CWLOutputType,
